Fixed typos and auto-save locations.
Apparently, the default `auto-save-file-name-transforms' action is to
only save *remote* files. Since it is is a list however, it is
appropriate to *add* a catch-all entry for all other files as well.

Also worth noting here: the .#....# files are *not* auto-save files, but
lock-files and are controlled by a separate mechanism and should be left alone.