Updated doxy yasnippet.
The `forward-thing` and `semantic-current-tag` functions were rather
fragile in some cases. Especially inside classes or namespaces. However,
using *senator* it quickly and accurately finds the next function.

Also, dtors/ctors no longer creates a @return clause.

It is however not noticing the *class ...* and *namespace ...* locations
for a Doxygen entry, but we could create a separate snippet for those,
since they are considerably easier and less defined that functions.