Scripts: Bugfixes for `screen_gif.sh`.
With large files, `imagemagick` would hand over processing to ffmpeg to
convert the video to a GIF. However, the command it used turned out to
be outdated, which caused the program to crash.

This fix removes `imagemagick` altogether, replacing it with a few
optimized ffmpeg commands.

Also, should the recording fail, the temporary files will remain.